requirements.txt
# Procesamiento XML y datos
lxml==4.9.4
xmltodict==0.13.0
python-dotenv==1.0.0

# Entorno virtual
virtualenv==20.25.1

# Backend/Embeddings
langchain==0.1.16
faiss-cpu==1.8.0
pymilvus==2.4.0

# Modelo LLM
transformers==4.40.0
torch==2.2.1
accelerate==0.29.3

# Frontend
streamlit==1.33.0

# Backend
fastapi==0.115.11
uvicorn==0.34.0
requests==2.32.3

# Adicionales
sentence-transformers==3.4.1
pandas==2.2.3